requirements.txt
django==1.8.4
djangorestframework==3.2.3
drf-nested-routers==1.1.2
django-flat-theme==1.1.1
bootstrapform==3.2
celery==3.1.18
